Put tests in the right place
